refactor: split test_arch_compliance.py from test_script_hygiene.py
test_script_hygiene.py was 1597 lines — double the 800-line wall it
enforces on scripts. Split architecture compliance tests (rules 4, 5,
7, 9, phase separation, feather handoff) into test_arch_compliance.py.

test_script_hygiene.py (1216L): coding conventions (sys.path, argparse,
complexity, naming, typing, print, I/O discipline, archive invariance).

test_arch_compliance.py (407L): pipeline architecture rules (phase
separation, compute/plot separation, save_figure, seeds, loaders).

Drop section numbers from headers — they're meaningless after a split.
